move the screen classes into the console package.
