+++ src/main/kotlin/presentation/Main.kt
@@ -5,7 +5,7 @@ import org.example.di.appModule
 import org.example.di.mongoModule
 import org.example.di.uiModule
 import org.example.di.useCaseModule
-import org.example.presentation.PlanMateConsoleUI
+import presentation.console.PlanMateConsoleUI
 import org.koin.core.context.startKoin
 import org.koin.java.KoinJavaComponent.getKoin
