+++ src/main/kotlin/presentation/console/TaskEditScreen.kt
@@ -1,6 +1,5 @@
-package org.example.presentation
+package presentation.console
 
-import SwimlaneRenderer
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.datetime.Clock
